requirements.txt
aiogram>=3.0
requests
beautifulsoup4
lxml
python-dotenv
